Fix bad usages of mktime reported by PHPstan

+++ front/stat.location.php
@@ -62,8 +62,8 @@ foreach (['date1', 'date2'] as $key) {
     }
 }
 if (empty($_GET["date1"]) && empty($_GET["date2"])) {
-    $year          = date("Y") - 1;
-    $_GET["date1"] = date("Y-m-d", mktime(1, 0, 0, date("m"), date("d"), $year));
+    $year          = ((int) date("Y")) - 1;
+    $_GET["date1"] = date("Y-m-d", mktime(1, 0, 0, (int) date("m"), (int) date("d"), $year));
     $_GET["date2"] = date("Y-m-d");
 }
